Move grouping and subshell parser
