+++ yash-syntax/src/parser.rs
@@ -22,6 +22,7 @@ mod core;
 mod fill;
 mod fromstr;
 
+mod grouping;
 mod redir;
 mod simple_command;
 
@@ -45,68 +46,6 @@ pub use self::fill::Fill;
 pub use self::fill::MissingHereDoc;
 
 impl Parser<'_> {
-    /// Parses a normal grouping.
-    ///
-    /// The next token must be a `{`.
-    ///
-    /// # Panics
-    ///
-    /// If the first token is not a `{`.
-    async fn grouping(&mut self) -> Result<CompoundCommand<MissingHereDoc>> {
-        let open = self.take_token_raw().await?;
-        assert_eq!(open.id, Token(Some(OpenBrace)));
-
-        let list = self.maybe_compound_list_boxed().await?;
-
-        let close = self.take_token_raw().await?;
-        if close.id != Token(Some(CloseBrace)) {
-            let opening_location = open.word.location;
-            let cause = SyntaxError::UnclosedGrouping { opening_location }.into();
-            let location = close.word.location;
-            return Err(Error { cause, location });
-        }
-
-        // TODO allow empty subshell if not POSIXly-correct
-        if list.0.is_empty() {
-            let cause = SyntaxError::EmptyGrouping.into();
-            let location = open.word.location;
-            return Err(Error { cause, location });
-        }
-
-        Ok(CompoundCommand::Grouping(list))
-    }
-
-    /// Parses a subshell.
-    ///
-    /// The next token must be a `(`.
-    ///
-    /// # Panics
-    ///
-    /// If the first token is not a `(`.
-    async fn subshell(&mut self) -> Result<CompoundCommand<MissingHereDoc>> {
-        let open = self.take_token_raw().await?;
-        assert_eq!(open.id, Operator(OpenParen));
-
-        let list = self.maybe_compound_list_boxed().await?;
-
-        let close = self.take_token_raw().await?;
-        if close.id != Operator(CloseParen) {
-            let opening_location = open.word.location;
-            let cause = SyntaxError::UnclosedSubshell { opening_location }.into();
-            let location = close.word.location;
-            return Err(Error { cause, location });
-        }
-
-        // TODO allow empty subshell if not POSIXly-correct
-        if list.0.is_empty() {
-            let cause = SyntaxError::EmptySubshell.into();
-            let location = open.word.location;
-            return Err(Error { cause, location });
-        }
-
-        Ok(CompoundCommand::Subshell(list))
-    }
-
     /// Parses a `do` clause, i.e., a compound list surrounded in `do ... done`.
     ///
     /// Returns `Ok(None)` if the first token is not `do`.
@@ -824,162 +763,6 @@ mod tests {
     use crate::alias::{AliasSet, HashEntry};
     use crate::source::{Location, Source};
     use futures::executor::block_on;
-
-    #[test]
-    fn parser_grouping_short() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "{ :; }");
-        let mut parser = Parser::new(&mut lexer);
-
-        let result = block_on(parser.compound_command()).unwrap().unwrap();
-        let result = result.fill(&mut std::iter::empty()).unwrap();
-        if let CompoundCommand::Grouping(list) = result {
-            assert_eq!(list.to_string(), ":");
-        } else {
-            panic!("Not a grouping: {:?}", result);
-        }
-    }
-
-    #[test]
-    fn parser_grouping_long() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "{ foo; bar& }");
-        let mut parser = Parser::new(&mut lexer);
-
-        let result = block_on(parser.compound_command()).unwrap().unwrap();
-        let result = result.fill(&mut std::iter::empty()).unwrap();
-        if let CompoundCommand::Grouping(list) = result {
-            assert_eq!(list.to_string(), "foo; bar&");
-        } else {
-            panic!("Not a grouping: {:?}", result);
-        }
-    }
-
-    #[test]
-    fn parser_grouping_unclosed() {
-        let mut lexer = Lexer::with_source(Source::Unknown, " { oh no ");
-        let mut parser = Parser::new(&mut lexer);
-
-        let e = block_on(parser.compound_command()).unwrap_err();
-        if let ErrorCause::Syntax(SyntaxError::UnclosedGrouping { opening_location }) = e.cause {
-            assert_eq!(opening_location.line.value, " { oh no ");
-            assert_eq!(opening_location.line.number.get(), 1);
-            assert_eq!(opening_location.line.source, Source::Unknown);
-            assert_eq!(opening_location.column.get(), 2);
-        } else {
-            panic!("Wrong error cause: {:?}", e.cause);
-        }
-        assert_eq!(e.location.line.value, " { oh no ");
-        assert_eq!(e.location.line.number.get(), 1);
-        assert_eq!(e.location.line.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 10);
-    }
-
-    #[test]
-    fn parser_grouping_empty_posix() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "{ }");
-        let mut parser = Parser::new(&mut lexer);
-
-        let e = block_on(parser.compound_command()).unwrap_err();
-        assert_eq!(e.cause, ErrorCause::Syntax(SyntaxError::EmptyGrouping));
-        assert_eq!(e.location.line.value, "{ }");
-        assert_eq!(e.location.line.number.get(), 1);
-        assert_eq!(e.location.line.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 1);
-    }
-
-    #[test]
-    fn parser_grouping_aliasing() {
-        let mut lexer = Lexer::with_source(Source::Unknown, " { :; end ");
-        let mut aliases = AliasSet::new();
-        let origin = Location::dummy("".to_string());
-        aliases.insert(HashEntry::new(
-            "{".to_string(),
-            "".to_string(),
-            false,
-            origin.clone(),
-        ));
-        aliases.insert(HashEntry::new(
-            "}".to_string(),
-            "".to_string(),
-            false,
-            origin.clone(),
-        ));
-        aliases.insert(HashEntry::new(
-            "end".to_string(),
-            "}".to_string(),
-            false,
-            origin,
-        ));
-        let mut parser = Parser::with_aliases(&mut lexer, std::rc::Rc::new(aliases));
-
-        let result = block_on(parser.compound_command()).unwrap().unwrap();
-        let result = result.fill(&mut std::iter::empty()).unwrap();
-        if let CompoundCommand::Grouping(list) = result {
-            assert_eq!(list.to_string(), ":");
-        } else {
-            panic!("Not a grouping: {:?}", result);
-        }
-    }
-
-    #[test]
-    fn parser_subshell_short() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "(:)");
-        let mut parser = Parser::new(&mut lexer);
-
-        let result = block_on(parser.compound_command()).unwrap().unwrap();
-        let result = result.fill(&mut std::iter::empty()).unwrap();
-        if let CompoundCommand::Subshell(list) = result {
-            assert_eq!(list.to_string(), ":");
-        } else {
-            panic!("Not a subshell: {:?}", result);
-        }
-    }
-
-    #[test]
-    fn parser_subshell_long() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "( foo& bar; )");
-        let mut parser = Parser::new(&mut lexer);
-
-        let result = block_on(parser.compound_command()).unwrap().unwrap();
-        let result = result.fill(&mut std::iter::empty()).unwrap();
-        if let CompoundCommand::Subshell(list) = result {
-            assert_eq!(list.to_string(), "foo& bar");
-        } else {
-            panic!("Not a subshell: {:?}", result);
-        }
-    }
-
-    #[test]
-    fn parser_subshell_unclosed() {
-        let mut lexer = Lexer::with_source(Source::Unknown, " ( oh no");
-        let mut parser = Parser::new(&mut lexer);
-
-        let e = block_on(parser.compound_command()).unwrap_err();
-        if let ErrorCause::Syntax(SyntaxError::UnclosedSubshell { opening_location }) = e.cause {
-            assert_eq!(opening_location.line.value, " ( oh no");
-            assert_eq!(opening_location.line.number.get(), 1);
-            assert_eq!(opening_location.line.source, Source::Unknown);
-            assert_eq!(opening_location.column.get(), 2);
-        } else {
-            panic!("Wrong error cause: {:?}", e.cause);
-        }
-        assert_eq!(e.location.line.value, " ( oh no");
-        assert_eq!(e.location.line.number.get(), 1);
-        assert_eq!(e.location.line.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 9);
-    }
-
-    #[test]
-    fn parser_subshell_empty_posix() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "( )");
-        let mut parser = Parser::new(&mut lexer);
-
-        let e = block_on(parser.compound_command()).unwrap_err();
-        assert_eq!(e.cause, ErrorCause::Syntax(SyntaxError::EmptySubshell));
-        assert_eq!(e.location.line.value, "( )");
-        assert_eq!(e.location.line.number.get(), 1);
-        assert_eq!(e.location.line.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 1);
-    }
 
     #[test]
     fn parser_do_clause_none() {
